Relocate `Common::Logging` into `bosh-director`
- collapse `Loggin` and `RegexFilter` into `RegexLoggingFilter`
